updated loading
Right now the loading screen takes longer but it is much better this way
because we dont get a splash of the login page when the user is logged
in already. It is covered by the loading div. I do this by defiing a
variable isLoading on app-component and subscribe to auth.user . When
the result is returned i set isloading to false because now the view is
updated.